+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 # RxDB Changelog
 
 <!-- CHANGELOG NEWEST -->
-
+- ADD [Supabase Replication Plugin](https://rxdb.info/replication-supabase.html) (beta)
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->